+++ core/index.d.ts
@@ -59,6 +59,7 @@ export type {
 } from "./src/config/createLytxAppConfig";
 export type {
   LytxAdditionalRoute,
+  LytxDocumentComponent,
   LytxDashboardReportData,
   LytxDashboardRouteUiOverrideArgs,
   LytxEventsRouteUiOverrideArgs,